data flow structure for OTP, username|password auth

+++ IDPServer/Models/ApplicationRole.cs
@@ -2,7 +2,7 @@ using Microsoft.AspNetCore.Identity;
 
 namespace IDPServer.Models;
 
-public sealed class ApplicationRole : IdentityRole<int>
+public sealed class ApplicationRole : IdentityRole<long>
 {
     public ApplicationRole() : base() {}
     public ApplicationRole(string roleName) : base(roleName) {}
