+++ IDPServer/Models/ApplicationUserToken.cs
@@ -2,6 +2,6 @@ using Microsoft.AspNetCore.Identity;
 
 namespace IDPServer.Models;
 
-public sealed class ApplicationUserToken : IdentityUserToken<int>
+public sealed class ApplicationUserToken : IdentityUserToken<long>
 {
 }
